fix: Adjust logging levels for more detailed debugging
This commit adjusts the logging levels for improved debugging in the artist module. The ROOT and tech.jhipster log levels are changed to Debug, and the org.hibernate.SQL and com.artbridge.artist log levels are also set to Debug for more detailed information. Additionally, the org.apache.kafka.clients log level is set to off to reduce unnecessary logging. The com.netflix log level remains as info. This update enhances the logging configuration to facilitate better debugging and troubleshooting in the artist module.